Fix GPU build
I removed a lot of things that prevented the code from being built
on Gpus, e.g. vectors inside ParallelFor regions. I also added
AMReX's GPU macros in front of the functions that need to be
offloaded e.g. the calculation of the potential and initial
conditions.
I also turned off some of the Catalyst functionality for now.